Sparse poa fix

+++ include/pacbio/consensus/poa/RangeFinder.h
@@ -35,7 +35,7 @@ class PoaGraphImpl;
 class SdpRangeFinder
 {
 private:
-    std::map<PoaGraph::Vertex, std::pair<size_t, size_t>> alignableReadIntervalByVertex_;
+    std::map<PoaGraph::Vertex, std::pair<int, int>> alignableReadIntervalByVertex_;
 
 public:
     virtual ~SdpRangeFinder();
@@ -44,9 +44,13 @@ public:
                          const std::vector<PoaGraph::Vertex>& consensusPath,
                          const std::string& consensusSequence, const std::string& readSequence);
 
-    std::pair<size_t, size_t> FindAlignableRange(PoaGraph::Vertex v);
+
+    // TODO: write contract
+    std::pair<int, int> FindAlignableRange(PoaGraph::Vertex v);
 
 protected:
+
+    // TODO: write contract
     virtual SdpAnchorVector FindAnchors(const std::string& consensusSequence,
                                         const std::string& readSequence) const = 0;
 };
